fix gem loading situation.
