+++ database_cleaner-core.gemspec
@@ -14,8 +14,8 @@ Gem::Specification.new do |spec|
   spec.license     = "MIT"
 
   spec.files         = `git ls-files -z`.split("\x0").reject do |f|
-    f.match(%r{^(test|spec|features)/})
-  end
+    f.match(%r{^(test|spec|features|examples)/})
+  end - ["lib/database_cleaner.rb"] # should only exist in database_cleaner gem
   spec.bindir        = "exe"
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ["lib"]
@@ -27,4 +27,3 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "rspec"
   spec.add_development_dependency "cucumber"
 end
-
