+++ features/support/env.rb
@@ -1,4 +1,4 @@
 $LOAD_PATH.unshift(File.dirname(__FILE__) + '/../../lib')
-require 'database_cleaner'
+require 'database_cleaner-core'
 
 require 'rspec/expectations'
