+++ lib/database_cleaner-core.rb
@@ -1,1 +1,1 @@
-# empty file to avoid LoadError if someone requires database_cleaner-core directly
+require "database_cleaner/core"
